move alle exceptions to errors
