+++ services/api-server/src/simcore_service_api_server/api/routes/solvers_jobs_getters.py
@@ -21,9 +21,9 @@ from servicelib.fastapi.requests_decorators import cancel_on_disconnect
 from servicelib.logging_utils import log_context
 from starlette.background import BackgroundTask
 
+from ...errors.custom_exceptions import InsufficientCreditsError, MissingWalletError
 from ...errors.services_exception_handlers import DEFAULT_BACKEND_SERVICE_STATUS_CODES
 from ...models.basic_types import LogStreamingResponse, VersionStr
-from ...models.custom_errors import InsufficientCreditsError, MissingWalletError
 from ...models.pagination import Page, PaginationParams
 from ...models.schemas.errors import ErrorGet
 from ...models.schemas.files import File
