+++ services/api-server/src/simcore_service_api_server/errors/_custom_errors_handlers.py
@@ -1,7 +1,7 @@
 from fastapi import Request, status
 from starlette.responses import JSONResponse
 
-from ..models.custom_errors import (
+from .custom_exceptions import (
     CustomBaseError,
     InsufficientCreditsError,
     MissingWalletError,
