+++ services/api-server/src/simcore_service_api_server/errors/base_exceptions.py
