+++ services/api-server/src/simcore_service_api_server/errors/custom_exceptions.py
@@ -1,4 +1,4 @@
-from .errors import ApiServerBaseError
+from .base_exceptions import ApiServerBaseError
 
 
 class CustomBaseError(ApiServerBaseError):
